Improved hill-climbing heuristic
Hill-climbing heuristic improved using a predom value.

Due to the method of changing one allele at a time, the use of sigCoefficient can result in a lack of optimization in hill-climbing.  Predom does not take sigCoefficient into account so that hill-climbing can choose new solutions that are more optimal, but not significantly more optimal. Dominance is still recorded so it can be compared against the genetic algorithm.

Genetic algorithm still consistently outperforms hill-climbing in optimization. Hill-climbing is still faster than the genetic algorithm.